Move graph building to navmesh.ts
